3：Kitbag/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-os-kitbag"
version = "1.0.0"
description = "Tool execution hub for Agent OS - declarative YAML-based tool registration and multi-protocol execution"
authors = [{name = "Agent OS Team"}]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
keywords = ["agent", "tools", "execution", "yaml", "multi-protocol"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "agent-os-common>=1.0.0",
    "PyYAML>=6.0",
    "httpx>=0.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/agent-os/agent-os"
Documentation = "https://agent-os.readthedocs.io"
Repository = "https://github.com/agent-os/agent-os"

[tool.setuptools.packages.find]
where = ["."]
include = ["agent_os.kitbag*"]
namespaces = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = "-v --tb=short --strict-markers"

[tool.black]
line-length = 100
target-version = ["py310", "py311"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "W", "I", "N"]
ignore = ["E501"]